+++ test/docker_tests/s3_test_data/simple_test.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+echo success!
+
+which python
+
+which pip
+
+which conda