+++ test/test_suite.py
@@ -1668,57 +1668,57 @@ class test_readcomparer():
         # # # ### self.tearDown()
 
 
-        # self.setUp()
-        # self.UPDATE_COMPARE_TEST_DATA()
-        # self.tearDown()
+        self.setUp()
+        self.UPDATE_COMPARE_TEST_DATA()
+        self.tearDown()
 
-        # self.setUp()
-        # self.testS()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test0()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test1()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test2()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test3()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test4()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test5()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test6()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test7()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test8()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test11()
-        # self.tearDown()
-        #
-        # self.setUp()
-        # self.test12()
-        # self.tearDown()
+        self.setUp()
+        self.testS()
+        self.tearDown()
+
+        self.setUp()
+        self.test0()
+        self.tearDown()
+
+        self.setUp()
+        self.test1()
+        self.tearDown()
+
+        self.setUp()
+        self.test2()
+        self.tearDown()
+
+        self.setUp()
+        self.test3()
+        self.tearDown()
+
+        self.setUp()
+        self.test4()
+        self.tearDown()
+
+        self.setUp()
+        self.test5()
+        self.tearDown()
+
+        self.setUp()
+        self.test6()
+        self.tearDown()
+
+        self.setUp()
+        self.test7()
+        self.tearDown()
+
+        self.setUp()
+        self.test8()
+        self.tearDown()
+
+        self.setUp()
+        self.test11()
+        self.tearDown()
+
+        self.setUp()
+        self.test12()
+        self.tearDown()
 
         self.setUp(destroy=True)
         self.test13()
@@ -4261,14 +4261,14 @@ class test_special():
 
 
 if __name__ == '__main__':
-    # test_strains().run()
-    # test_filter_reads().run()
-    # test_SNVprofile().run()
-    # test_gene_statistics().run()
-    # test_quickProfile().run()
-    # test_genome_wide().run()
-    # test_plot().run()
-    # test_special().run()
+    test_strains().run()
+    test_filter_reads().run()
+    test_SNVprofile().run()
+    test_gene_statistics().run()
+    test_quickProfile().run()
+    test_genome_wide().run()
+    test_plot().run()
+    test_special().run()
     test_readcomparer().run()
 
     print('everything is working swimmingly!')